Struct tract_tensorflow::prelude::InferenceFact

pub struct InferenceFact {
    pub datum_type: GenericFactoid<DatumType>,
    pub shape: ShapeFactoid,
    pub value: GenericFactoid<Arc<Tensor>>,
}

Partial information about a tensor.

The task of the analyser is to tag every edge in the graph with information about the tensors that flow through it - specifically their datum_type, their shape and possibly their value. During the analysis, however, we might only know some of that information (say, for instance, that an edge only carries tensors of rank 4, but without knowing their precise dimension).

This is where tensor facts come in: they hold partial information about the datum_type, shape and value of tensors that might flow through an edge of the graph. The analyser will first tag each edge with a fact, starting with the most general one and specializing it at each iteration. Eventually, it will reach a fixed point that - hopefully - holds enough information.

Fields
datum_type: GenericFactoid<DatumType>shape: ShapeFactoidvalue: GenericFactoid<Arc<Tensor>>
Implementations

impl InferenceFact

pub fn new() -> InferenceFact

Constructs the most general tensor fact possible.

pub fn any() -> InferenceFact

pub fn dt(dt: DatumType) -> InferenceFact

pub fn dt_shape<S>(dt: DatumType, shape: S) -> InferenceFact where
S: Into<ShapeFactoid>,

pub fn shape<S>(shape: S) -> InferenceFact where
S: Into<ShapeFactoid>,

pub fn with_datum_type(self, dt: DatumType) -> InferenceFact

pub fn with_shape<S>(self, shape: S) -> InferenceFact where
S: Into<ShapeFactoid>,

pub fn format_dt_shape(&self) -> String

pub fn dt_shape_from_tensor(t: &Tensor) -> InferenceFact

pub fn without_value(self) -> InferenceFact